Added test case for MCXAdapter.
